+++ SharedLibrary/Events/IskUpdate.cs
@@ -1,0 +1,5 @@
+using SharedLibrary.Models;
+
+namespace SharedLibrary.Events;
+
+public record IskUpdate(int TotalBounty, int LastBounty, Character Character, int CharacterBounty);